// src/models/Variant.ts
export enum ColorVariant {
    red = "red",
    orange = "orange",
    yellow = "yellow",
    green = "green",
    sky = "sky",
    blue = "blue",
    purple = "purple",
    pink = "pink",
}

export enum ButtonVariant {
    inactive = "inactive",
    primary = "primary",
    secondary = "secondary",
    danger = "danger",
    accept = "accept",
}